+++ packages/server/base.js
@@ -1,4 +1,0 @@
-import path from 'path';
-import { fileURLToPath } from 'url';
-
-export const base = path.dirname(fileURLToPath(import.meta.url));