+++ packages/server/routers/index.js
@@ -1,7 +1,0 @@
-import * as config from './router.config.js';
-import * as mod from './router.mod.js';
-
-export default {
-    mod,
-    config,
-};