+++ packages/server/src/context.ts
@@ -1,0 +1,38 @@
+import type { CreateFastifyContextOptions } from '@trpc/server/adapters/fastify';
+
+/**
+ * Inner context. Will always be available in your procedures, in contrast to the outer context.
+ *
+ * Also useful for:
+ * - testing, so you don't have to mock Next.js' `req`/`res`
+ * - tRPC's `createServerSideHelpers` where we don't have `req`/`res`
+ *
+ * @see https://trpc.io/docs/context#inner-and-outer-context
+ */
+export async function createContextInner(opts?: CreateInnerContextOptions) {
+  return {
+    prisma,
+    session: opts.session,
+  };
+}
+/**
+ * Outer context. Used in the routers and will e.g. bring `req` & `res` to the context as "not `undefined`".
+ *
+ * @see https://trpc.io/docs/context#inner-and-outer-context
+ */
+export async function createContext(opts: CreateFastifyContextOptions) {
+  const session = getSessionFromCookie(opts.req);
+  const contextInner = await createContextInner({ session });
+  return {
+    ...contextInner,
+    req: opts.req,
+    res: opts.res,
+  };
+}
+
+const helpers = createServerSideHelpers({
+    router: appRouter,
+    ctx: await createContext(),
+  });
+
+export type Context = Awaited<ReturnType<typeof createContextInner>>;