+++ packages/server/src/middlewares/loginProxy.ts
@@ -1,7 +1,7 @@
 import { createProxyMiddleware } from 'http-proxy-middleware';
 import zlib from 'zlib';
 
-export const saLoginProxy = createProxyMiddleware({
+export const loginProxy = createProxyMiddleware({
     target: 'http://account-co.61.com/',
     changeOrigin: true,
     selfHandleResponse: true,
@@ -31,9 +31,8 @@ export const saLoginProxy = createProxyMiddleware({
             const headers = { ...proxyRes.headers };
             delete headers['content-length'];
 
-            /** @type {Buffer[]} */
-            const chunks = [];
-            proxyRes.on('data', (chunk) => chunks.push(chunk));
+            const chunks: Buffer[] = [];
+            proxyRes.on('data', (chunk: Buffer) => chunks.push(chunk));
 
             proxyRes.on('end', () => {
                 /** @type {string | Buffer} */
@@ -49,13 +48,13 @@ export const saLoginProxy = createProxyMiddleware({
                     body = body
                         .replaceAll(/: "\/v3/g, `: "/account-co.61.com/v3`)
                         .replaceAll(/"http:\/\/account-co.61.com/g, `"/account-co.61.com`);
-                    rawBuf = body;
+                    rawBuf = Buffer.from(body);
                 }
                 if (proxyRes.headers['content-encoding'] === 'gzip') {
                     rawBuf = zlib.gzipSync(rawBuf);
                 }
 
-                res.writeHead(proxyRes.statusCode ?? 200,headers);
+                res.writeHead(proxyRes.statusCode ?? 200, headers);
                 res.end(rawBuf);
             });
         },