+++ packages/server/src/router/routers/config.ts
@@ -1,34 +1,33 @@
+/* eslint-disable @typescript-eslint/no-unsafe-argument */
+/* eslint-disable @typescript-eslint/no-unsafe-return */
+/* eslint-disable @typescript-eslint/no-explicit-any */
+/* eslint-disable @typescript-eslint/no-unsafe-assignment */
 import fs from 'fs';
 import path from 'path';
 import * as toml from 'smol-toml';
 
-import { base } from '../base.js';
+import { basePath } from '../../config.ts';
 
 export const PetCache = {
-    configPath: path.join(base, 'config', 'sea-pet.toml'),
+    configPath: path.join(basePath, 'config', 'sea-pet.toml'),
     catchTimeMap: new Map(),
     load() {
         if (!fs.existsSync(this.configPath)) {
             fs.writeFileSync(this.configPath, '');
         }
         const content = fs.readFileSync(this.configPath, 'utf-8');
-        /** @type {any} */
-        const config = toml.parse(content);
+
+        const config: any = toml.parse(content);
         Object.keys(config).forEach((key) => this.catchTimeMap.set(key, config[key]));
     },
-    /**
-     * @param {string} name
-     */
-    query(name) {
+
+    query(name: string) {
         return this.catchTimeMap.get(name);
     },
-    /**
-     * @param {Map<string, number>} newData
-     */
-    update(newData) {
+
+    update(newData: Map<string, number>) {
         this.catchTimeMap = newData;
-        /** @type {Record<string, number>} */
-        const cacheObj = {};
+        const cacheObj: Record<string, number> = {};
         for (const [key, value] of newData.entries()) {
             cacheObj[key] = value;
         }
@@ -40,16 +39,13 @@ export const PetCache = {
 PetCache.load();
 
 export class ModConfig {
-    /** @type { Record<string, Record<string, object>> | undefined } */
-    data;
+    data: Record<string, Record<string, object>> | undefined;
 
     type;
     id;
     author;
-    /**
-     * @param {string} ns
-     */
-    constructor(ns) {
+
+    constructor(ns: string) {
         const { type, id, author } = praseNamespace(ns);
         this.type = type;
         this.id = id;
@@ -58,8 +54,7 @@ export class ModConfig {
         const configFile = configFilePath(author);
         if (fs.existsSync(configFile)) {
             const content = fs.readFileSync(configFile, 'utf-8');
-            /** @type {any} */
-            const config = toml.parse(content);
+            const config: any = toml.parse(content);
             this.data = config;
         }
     }
@@ -68,10 +63,7 @@ export class ModConfig {
         return this.data?.[this.type]?.[this.id];
     }
 
-    /**
-     * @param { Record<string, any> } data
-     */
-    save(data) {
+    save(data: Record<string, any>) {
         this.data = {
             ...this.data,
             [this.type]: {
@@ -87,11 +79,7 @@ export class ModConfig {
     }
 }
 
-/**
- * @param {string} ns
- * @returns {{type: string, author: string, id: string}}
- */
-function praseNamespace(ns) {
+function praseNamespace(ns: string): { type: string; author: string; id: string } {
     const [type, author, id] = ns.split('::');
     return {
         type,
@@ -100,9 +88,6 @@ function praseNamespace(ns) {
     };
 }
 
-/**
- * @param {string} filename
- */
-function configFilePath(filename) {
-    return path.join(base, 'config', `${filename}.toml`);
+function configFilePath(filename: string) {
+    return path.join(basePath, 'config', `${filename}.toml`);
 }